Added booking server side validation and modularized search and booking functionality.
